Реализовал файл с константами игры, и заменил double на decimal там, где используются деньги

+++ AquaManager/AquaManager.Domain/Interfaces/IFish.cs
@@ -29,7 +29,7 @@ namespace AquaManager.Domain.Interfaces
         /// <summary>
         /// Стоимость рыбки
         /// </summary>
-        public double Price { get; }
+        public decimal Price { get; }
 
         /// <summary>
         /// Рыбка жива
